requirements.txt
uvicorn
fastapi
torch
ffmpeg-python